tidy up code. move generate_urn_dataframe function into the change_item_topics.py file
