--- examples/generate_urn_dataframe.py
@@ -1,47 +1,0 @@
-from .lib.utility import get_urn_from_item, get_topic_of_item
-#pip install openpyxl #might need to install openpyxl, a dependency for read-excel
-import pandas as pd
-   
-def generate_urn_dataframe(input_file_name, C):
-    """Method for generating input for change_item_topics.update_topics. The code iterates through 
-    a spreadsheet containing details of new item topic assignments and generates a dataframe
-    of URNs that can be used as input to a method defined in change_item_topics.py that reassigns 
-    items to new topics. 
-    """
-    print(f"Reading topic reassignments from {input_file_name}")
-    data = pd.read_excel(input_file_name)
-    urn_data_frame={
-        "itemUrns": [],
-        "sourceTopicGroups": [],
-        "destinationTopicGroups": []
-    }
-    # Iterate through the rows in the spreadsheet. Each row contains details of a topic
-    # reassignment for an item...
-    for topic_reassignment_details in data.iloc:
-        instrument_name = topic_reassignment_details.iloc[0]
-        url = topic_reassignment_details.iloc[2]
-        agency_id = url.split("/")[4]
-        identifier = url.split("/")[5]
-        if len(url.split("/")) == 7:
-            version = url.split("/")[6]
-            item = C.get_item_json(agency_id, identifier, version=version)
-        else:
-            item = C.get_item_json(agency_id, identifier)
-        version = item['Version']
-        item_urn = "urn:ddi:" + agency_id + ":" + identifier + ":" + str(version)
-        item_type = item['ItemType']
-        if item_type==C.item_code('Question'):
-            topic_type=C.item_code('Question Group')
-            containing_item_type=C.item_code('Data Collection')
-        elif item_type==C.item_code('Variable'):
-            topic_type=C.item_code('Variable Group')
-            containing_item_type=C.item_code('Data File')
-        item_urn = get_urn_from_item(item)    
-        source_topic = get_topic_of_item(topic_reassignment_details.iloc[4], topic_type, instrument_name, containing_item_type, C)
-        destination_topic = get_topic_of_item(topic_reassignment_details.iloc[5], topic_type, instrument_name, containing_item_type, C)
-        urn_data_frame['itemUrns'].append(item_urn)
-        if len(source_topic)>0:
-            urn_data_frame['sourceTopicGroups'].append(get_urn_from_item(source_topic[0]))
-        if len(destination_topic)>0:
-            urn_data_frame['destinationTopicGroups'].append(get_urn_from_item(destination_topic[0])) 
-    return pd.DataFrame(urn_data_frame)
